parser: split AST type defs into ast.go
Pure code motion. parser.go now holds the lexer, the Parser instance,
ParseString, and AliasFromPath (108 lines). ast.go holds every AST
type def and its Capture/Name helpers (630 lines). Closes CLEAN-8.
